feat: Go back to start now on restart & Skip past demos
